# docs/openapi.yaml
openapi: 3.0.3
info:
  title: User Management API
  version: 1.0.0
  description: >
    API RESTful para gestión de usuarios y despliegues con autenticación JWT y
    control de acceso RBAC. Permite administrar usuarios, registrar despliegues
    de aplicaciones y consultar métricas de rendimiento.
  contact:
    name: Equipo de Desarrollo
    email: soporte@empresa.com

servers:
  - url: http://localhost:3000
    description: Servidor de desarrollo local
  - url: /
    description: Servidor de producción relativo

tags:
  - name: Auth
    description: Operaciones de autenticación y emisión de tokens JWT
  - name: Users
    description: Gestión CRUD de usuarios, requiere rol admin
  - name: Deployments
    description: Registro y consulta de despliegues de aplicaciones
  - name: Metrics
    description: Consulta de métricas de rendimiento por aplicación
  - name: Health
    description: Verificación del estado del servicio

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    UserBase:
      type: object
      required:
        - username
        - email
        - role
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 50
          description: Nombre de usuario único
        email:
          type: string
          format: email
          description: Dirección de correo electrónico del usuario
        role:
          type: string
          enum:
            - admin
            - viewer
          description: Rol del usuario en el sistema

    UserCreate:
      allOf:
        - $ref: '#/components/schemas/UserBase'
        - type: object
          required:
            - password
          properties:
            password:
              type: string
              minLength: 8
              format: password
              description: Contraseña del usuario (mínimo 8 caracteres)

    UserResponse:
      type: object
      properties:
        id:
          type: integer
          description: Identificador único del usuario
        username:
          type: string
          description: Nombre de usuario
        email:
          type: string
          description: Dirección de correo electrónico
        role:
          type: string
          description: Rol del usuario en el sistema
        created_at:
          type: string
          format: date-time
          description: Fecha y hora de creación del usuario

    UserUpdate:
      type: object
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 50
          description: Nombre de usuario único
        email:
          type: string
          format: email
          description: Dirección de correo electrónico del usuario
        role:
          type: string
          enum:
            - admin
            - viewer
          description: Rol del usuario en el sistema
        password:
          type: string
          minLength: 8
          description: Nueva contraseña del usuario (mínimo 8 caracteres)

    LoginRequest:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
          description: Nombre de usuario registrado
        password:
          type: string
          format: password
          description: Contraseña del usuario

    LoginResponse:
      type: object
      properties:
        token:
          type: string
          description: JWT firmado con HS256
        expires_in:
          type: integer
          description: Segundos hasta expiración del token

    DeploymentCreate:
      type: object
      required:
        - id_app
        - status
      properties:
        id_app:
          type: string
          description: Identificador único de la aplicación desplegada
        status:
          type: string
          enum:
            - SUCCESS
            - FAILED
            - IN_PROGRESS
          description: Estado del despliegue
        duration_seconds:
          type: integer
          minimum: 0
          description: Duración del despliegue en segundos
        deployed_by:
          type: string
          description: Usuario o sistema que realizó el despliegue

    DeploymentResponse:
      type: object
      properties:
        id:
          type: integer
          description: Identificador único del despliegue
        id_app:
          type: string
          description: Identificador único de la aplicación
        status:
          type: string
          description: Estado del despliegue
        duration_seconds:
          type: integer
          description: Duración del despliegue en segundos
        deployed_by:
          type: string
          description: Usuario o sistema que realizó el despliegue
        created_at:
          type: string
          format: date-time
          description: Fecha y hora en que se registró el despliegue

    MetricsResponse:
      type: object
      properties:
        id_app:
          type: string
          description: Identificador único de la aplicación
        average_duration_seconds:
          type: number
          format: float
          description: Tiempo promedio de despliegue exitoso en segundos
        total_deployments:
          type: integer
          description: Total de despliegues exitosos en el período
        period_days:
          type: integer
          example: 7
          description: Número de días del período analizado

    HealthResponse:
      type: object
      properties:
        status:
          type: string
          example: ok
          description: Estado general del servicio
        uptime:
          type: number
          description: Tiempo en segundos que el servicio lleva en ejecución
        timestamp:
          type: string
          format: date-time
          description: Fecha y hora actual del servidor

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Código o tipo de error
        message:
          type: string
          description: Descripción detallada del error

    PaginatedUsers:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/UserResponse'
          description: Lista de usuarios en la página actual
        total:
          type: integer
          description: Total de usuarios en el sistema
        page:
          type: integer
          description: Número de página actual
        limit:
          type: integer
          description: Cantidad de registros por página

  parameters:
    UserIdParam:
      name: id
      in: path
      required: true
      description: ID único del usuario
      schema:
        type: integer

    AppIdParam:
      name: id_app
      in: path
      required: true
      description: Identificador único de la aplicación
      schema:
        type: string

  responses:
    UnauthorizedError:
      description: Token JWT ausente o inválido
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'

    ForbiddenError:
      description: El usuario no tiene el rol requerido (admin)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'

    NotFoundError:
      description: Recurso no encontrado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'

    ValidationError:
      description: Error de validación en el payload
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'

    InternalServerError:
      description: Error interno del servidor
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'

paths:
  /api/v1/auth/login:
    post:
      tags:
        - Auth
      summary: Autenticar usuario y obtener JWT
      description: >
        Recibe credenciales de usuario y retorna un token JWT firmado con HS256
        para ser utilizado en los endpoints protegidos mediante el header
        Authorization Bearer.
      operationId: loginUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            example:
              username: admin
              password: secretpassword
      responses:
        '200':
          description: Autenticación exitosa, retorna token JWT
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
              example:
                token: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
                expires_in: 3600
        '401':
          description: Credenciales inválidas
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: INVALID_CREDENTIALS
                message: Usuario o contraseña incorrectos
        '422':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /api/v1/users:
    post:
      tags:
        - Users
      summary: Crear nuevo usuario
      description: >
        Crea un nuevo usuario en el sistema. Requiere autenticación JWT y rol
        admin. La contraseña se almacena hasheada con bcrypt.
      operationId: createUser
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreate'
            example:
              username: johndoe
              email: johndoe@empresa.com
              role: viewer
              password: securepass123
      responses:
        '201':
          description: Usuario creado exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
              example:
                id: 1
                username: johndoe
                email: johndoe@empresa.com
                role: viewer
                created_at: '2024-01-15T10:30:00Z'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '422':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalServerError'

    get:
      tags:
        - Users
      summary: Listar todos los usuarios
      description: >
        Retorna una lista paginada de todos los usuarios registrados en el
        sistema. Requiere autenticación JWT y rol admin.
      operationId: listUsers
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          required: false
          description: Número de página (por defecto 1)
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          required: false
          description: Cantidad de registros por página (por defecto 10)
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
      responses:
        '200':
          description: Lista de usuarios obtenida exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedUsers'
              example:
                data:
                  - id: 1
                    username: johndoe
                    email: johndoe@empresa.com
                    role: viewer
                    created_at: '2024-01-15T10:30:00Z'
                total: 1
                page: 1
                limit: 10
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /api/v1/users/{id}:
    get:
      tags:
        - Users
      summary: Obtener usuario por ID
      description: >
        Retorna los datos de un usuario específico identificado por su ID.
        Requiere autenticación JWT y rol admin.
      operationId: getUserById
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/UserIdParam'
      responses:
        '200':
          description: Usuario encontrado exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
              example:
                id: 1
                username: johndoe
                email: johndoe@empresa.com
                role: viewer
                created_at: '2024-01-15T10:30:00Z'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalServerError'

    put:
      tags:
        - Users
      summary: Actualizar usuario por ID
      description: >
        Actualiza los datos de un usuario existente. Todos los campos son
        opcionales. Requiere autenticación JWT y rol admin.
      operationId: updateUser
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/UserIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdate'
            example:
              email: newemail@empresa.com
              role: admin
      responses:
        '200':
          description: Usuario actualizado exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
              example:
                id: 1
                username: johndoe
                email: newemail@empresa.com
                role: admin
                created_at: '2024-01-15T10:30:00Z'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '422':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalServerError'

    delete:
      tags:
        - Users
      summary: Eliminar usuario por ID
      description: >
        Elimina permanentemente un usuario del sistema. Requiere autenticación
        JWT y rol admin.
      operationId: deleteUser
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/UserIdParam'
      responses:
        '204':
          description: Usuario eliminado exitosamente, sin contenido en la respuesta
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /api/v1/deployments:
    post:
      tags:
        - Deployments
      summary: Registrar nuevo despliegue
      description: >
        Registra un nuevo evento de despliegue de aplicación en el sistema.
        Requiere autenticación JWT y rol admin.
      operationId: createDeployment
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeploymentCreate'
            example:
              id_app: app-frontend-v2
              status: SUCCESS
              duration_seconds: 120
              deployed_by: johndoe
      responses:
        '201':
          description: Despliegue registrado exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeploymentResponse'
              example:
                id: 1
                id_app: app-frontend-v2
                status: SUCCESS
                duration_seconds: 120
                deployed_by: johndoe
                created_at: '2024-01-15T10:30:00Z'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '422':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /api/v1/metrics/average-time/{id_app}:
    get:
      tags:
        - Metrics
      summary: Obtener tiempo promedio de despliegues exitosos
      description: >
        Retorna el tiempo promedio de duración de los despliegues con estado
        SUCCESS para una aplicación específica durante los últimos 7 días.
        Requiere autenticación JWT y rol admin.
      operationId: getAverageDeploymentTime
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/AppIdParam'
      responses:
        '200':
          description: Métricas obtenidas exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MetricsResponse'
              example:
                id_app: app-frontend-v2
                average_duration_seconds: 95.5
                total_deployments: 12
                period_days: 7
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /api/v1/sync/users:
    post:
      tags:
        - Users
      summary: Sincronizar usuarios con servicios externos
      description: >
        Dispara la sincronización de usuarios con servicios externos mediante
        mensajería asíncrona (RabbitMQ). Requiere autenticación JWT y rol admin.
      operationId: syncUsers
      security:
        - BearerAuth: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                source:
                  type: string
                  description: Fuente externa desde la cual sincronizar usuarios
                force:
                  type: boolean
                  description: Forzar sincronización completa ignorando caché
                  default: false
            example:
              source: ldap
              force: false
      responses:
        '202':
          description: Solicitud de sincronización aceptada y encolada
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    description: Mensaje de confirmación
                  job_id:
                    type: string
                    description: Identificador del trabajo de sincronización encolado
              example:
                message: Sincronización encolada exitosamente
                job_id: sync-1705312200-abc123
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /api/health:
    get:
      tags:
        - Health
      summary: Verificar estado del servicio
      description: >
        Retorna el estado actual del servicio, tiempo de actividad y timestamp
        del servidor. No requiere autenticación.
      operationId: healthCheck
      responses:
        '200':
          description: El servicio está operativo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResponse'
              example:
                status: ok
                uptime: 3600.5
                timestamp: '2024-01-15T10:30:00Z'
        '500':
          $ref: '#/components/responses/InternalServerError'